refactor(utils): relocate Subrect into Util namespace
Move src/Utils/Subrect/Subrect.{h,cpp} to src/Utils/Subrect.{h,cpp}
to match the flat layout of other Utils helpers (FileSystem, D3D, UI),
and rename `namespace Subrect` to `namespace Util::Subrect` so it
no longer reads like a standalone module.

Why: per #2092 review, helpers under src/Utils should follow the
established Util:: convention rather than introducing their own
top-level namespace and subdirectory.
